v3.0.0 删除view，加入路由和Dockerfile
