+++ app/controller/index.class.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace App\Controller;
+use Core\Base\Controller;
+
+class IndexController extends Controller
+{
+    public function index()
+    {
+        $this->jsonReturn([
+            'status' => 200,
+            'body' => C('body')
+        ]);
+    }
+}
