+++ config/extend/config.ini.php
@@ -1,0 +1,1 @@
+<?php//平台扩展配置return [    'body' => '欢迎使用SSF(Simple-Swoole-Framework)! 您的应用已经部署成功!'];
