+++ config/system/database.ini.php
@@ -1,0 +1,18 @@
+<?php
+/**
+ * 数据库配置文件
+ * "DEFAULT"键名为默认配置，可在不特指的情况下直接使用
+ */
+return [
+//    'DEFAULT' => [
+//        'TYPE'   => 'mysql', // 数据库类型
+//        'HOST'   => '127.0.0.1', // 服务器地址
+//        'NAME'   => 'test', // 数据库名
+//        'USER'   => 'root', // 用户名
+//        'PASS'   => '', // 密码
+//        'PORT'   => '3306', // 端口
+//        'PREFIX' => '', // 数据库表前缀
+//        'CHARSET'=> 'utf8mb4', // 字符集
+//        'POOL'   => 0  //连接池最大容量(0为禁用连接池，每次会话结束后连接将会被回收)，连接池中的连接只有在需要的时候才会创建，如果5分钟没有使用则会被回收
+//    ]
+];
