+++ config/system/map.ini.php
@@ -1,0 +1,8 @@
+<?php
+return [
+    //加载类型 0-按目录文件自动加载 1-率先按顺序加载表中的文件 2-只按顺序加载表中的文件
+    'TYPE' => 0,
+    //手动加载表，填写文件在app文件夹中的相对路径，如"common/main.class.php"
+    'LIST' => [
+    ]
+];
