+++ config/system/mtable.ini.php
@@ -1,0 +1,27 @@
+<?php
+/**
+ * 内存表配置
+ * 内存表是在服务启动之前创建的，所有工作进程及子进程都可以使用
+ */
+return [
+//    'mytable' => [
+//        '__total' => 10, //内存表条目总数（2^10=1024）
+//        '__expire' => 1 * 60 * 60, //条目有效期一小时
+//        'id' => [
+//            'type' => 'int', //字段类型
+//            'size' => '4', //字段占用大小
+//            'auto' => 'increment' //自动递增
+//        ],
+//        'data' => 'string(40)', //字符串的设置方式
+//        'create_time' => [
+//            'type' => 'int', //时间戳
+//            'size' => '4',
+//            'auto' => 'datetime' //自动取当前时间（示例：1610712027）
+//        ],
+//        'datetime' => [
+//            'type' => 'string', //字符串时间
+//            'size' => '19',
+//            'auto' => 'datetime' //自动取当前时间（示例：2021-01-15 12:00:27）
+//        ]
+//    ]
+];
