+++ config/system/process.ini.php
@@ -1,0 +1,8 @@
+<?php
+return [
+    //子进程配置
+//    'process_name' => [
+//        'worker_num' => 1, //子进程进程数量
+//        'class_name' => '\App\Common\Sub' //子进程实例类名
+//    ]
+];
