+++ config/system/redis.ini.php
@@ -1,0 +1,15 @@
+<?php
+/**
+ * REDIS客户端配置
+ * "DEFAULT"键名为默认配置，可在不特指的情况下直接使用
+ */
+return [
+//    'DEFAULT' => [
+//        'HOST' => '127.0.0.1', //redis连接IP
+//        'PORT' => '6379', //redis连接端口
+//        'AUTH' => '', //redis连接秘钥
+//        'DB' => 0,
+//        'EXPIRE' => 3600, //默认字段有效期
+//        'USE_COROUTINE' => false //是否使用协程redis客户端
+//    ]
+];
