+++ core/ini/mtable.ini.php
@@ -1,0 +1,15 @@
+<?php
+return [
+    '__SESSION' => [ //表名
+        '__total' => 10, //内存表总行数(2的指数), 2的10次方等于1024
+        '__expire' => 3 * 60 * 60, //有效期3小时
+        'timeout' => 'int(4)', //字段名
+        'data' => 'string(10240)' //字段名 长度最大10KB
+    ],
+    '__LOCK' => [
+        '__total' => 10,
+        '__expire' => 60, //有效期60秒
+        'type' => 'int(1)', //锁类型, 0-队列锁 1-异步锁
+        'timeout' => 'int(4)' //锁的失效时间，到时间会自动解锁
+    ]
+];
